Improve performance by reusing workspace and project reader
These two were per-build, which caused a significant amount of time to be spent spinning them up and resolving assemblies.

This new pattern of per-AppDomain registration, plus a per-Build cleanup object, significantly improves performance to the point of being almost instantaneous for a small project.

As a side-effect, the code fix providers can no longer be retrieved from the workspace itself (otherwise it couldn't be reusable) and we instead load them in the task itself. This had the additionally benefit that now each project can provide a different set of analyzers and code fixers which won't collide with each other.