Refactored initialization code to its own target/task
By moving this out of the ApplyCodeFixes task, that task receives fewer parameters and can rely on build-registered objects to be there already, simplying the implementation and centralizing the initialization in a single task.

This also allows that initialization to happen early on in the code, which specially for the project reader, means that by the time we actually need to read the project, the main load has already happened and we can provide a quicker run.