Add support for batch fixers
This improves performance massively when there are lots of diagnostics and
the code fixer supports fix all. "Fixing" SA1309 (remove underscore from
field names in the Microsoft.CodeAnalysis project) went from an overnight
thing (several hours) to: `Fixed 1541 diagnostics in 7 minutes`. Note that
this is a non-trivial code fix that does a rename refactoring actually, so
it's very much real-world.

In most cases, though, fixes take milliseconds, and for the main use case
of automatically fixing code as you introduce it, it's awesome.