Align version with Roslyn version
Since AutoCodeFix needs to align with Roslyn versions to
properly load analyzers and code fixers, we should version
our package accordingly, and build multiple versions for
each supported Roslyn version. This way, adding new stable
versions is as easy as adding one more templatized build
step to the pipeline. Over time we can drop older versions,
since the published packages can't be pushed again, since
our stable versions will match roslyn versions too.

The template receives two parameters: `roslyn` for the Roslyn
version to build for, and the optional `package`, which is the
version to produce for the AutoCodeFix package. This second
one will only be used if we're patching a previously shipped
version based on the roslyn one (i.e. 2.10.0), where the package
can be, for example, 2.10.1. This would allow consumers to
safely wildcard the patch segment to get minor fixes for stable
versions of AutoCodeFix against a specific version of Roslyn.