Fix bug in the flexible version resolving
If located local major version is greater  than the requested one, we should always succeed. Only then, additionally, support binding to a higher minor version if the major is the same.

This issue was causing Moq to not generate code when AutoCodeFix moved to 3.0.0, since the old check would check for *both* major and minor being greater or equal, which wasn't the case for Moq's 2.10.0 dependency.